+++ src/privacy/aes128.rs
@@ -1,11 +1,11 @@
 // ------------------------------------------------------------------------
 // Gufo SNMP: AES128 mode
 // ------------------------------------------------------------------------
-// Copyright (C) 2023-24, Gufo Labs
+// Copyright (C) 2023-25, Gufo Labs
 // See LICENSE.md for details
 // ------------------------------------------------------------------------
 
-use super::SnmpPriv;
+use super::{SnmpPriv, get_padded_len};
 use crate::ber::BerEncoder;
 use crate::buf::Buffer;
 use crate::error::{SnmpError, SnmpResult};
@@ -61,13 +61,7 @@ impl SnmpPriv for Aes128Key {
         // Serialize
         pdu.push_ber(&mut self.buf)?;
         // Calculate length
-        let scoped_pdu_len = self.buf.len() - BLOCK_SIZE;
-        let rem = scoped_pdu_len % BLOCK_SIZE;
-        let padded_len = if rem > 0 {
-            scoped_pdu_len + BLOCK_SIZE - rem
-        } else {
-            scoped_pdu_len
-        };
+        let padded_len = get_padded_len(self.buf.len() - BLOCK_SIZE, BLOCK_SIZE);
         // Encrypt
         let encryptor = Aes128CfbEncryptor::new_from_slices(&self.key, &self.priv_params)
             .map_err(|_| SnmpError::InvalidKey)?;