+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 To see unreleased changes, please see the [CHANGELOG on the main branch guide](https://github.com/gufolabs/gufo_snmp/blob/main/CHANGELOG.md).
 
+## [Upcoming]
+
+### Fixed
+
+* Fixed DES privacy padding in SNMP v3.
+
 ## 0.8.0 - 2025-04-15
 
 ### Added